Changed Box.TimeToGraph and Box.TimeLenght from integer to double

+++ src/BoxOptions.Services/Models/Box.cs
@@ -10,16 +10,16 @@ namespace BoxOptions.Services.Models
         string id;
         decimal minPrice;
         decimal maxPrice;
-        int timeToGraph; // (in seconds), 
-        int timeLength;//(in seconds), 
+        double timeToGraph; // (in seconds), 
+        double timeLength;//(in seconds), 
         decimal coefficient;
         
                 
         public string Id { get => id; set => id= value; }
         public decimal MinPrice { get => minPrice; set => minPrice = value; }
         public decimal MaxPrice { get => maxPrice; set => maxPrice = value; }
-        public int TimeToGraph { get => timeToGraph; set => timeToGraph = value; }
-        public int TimeLength { get => timeLength; set => timeLength = value; }
+        public double TimeToGraph { get => timeToGraph; set => timeToGraph = value; }
+        public double TimeLength { get => timeLength; set => timeLength = value; }
         public decimal Coefficient { get => coefficient; set => coefficient = value; }
                
         public static Box FromJson(string json)
